% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_auroc.R
\name{km_auroc}
\alias{km_auroc}
\title{Kaplan-Meier adjusted AUROC}
\usage{
km_auroc(evaluate_at, time, event, log_risk, data)
}
\arguments{
\item{evaluate_at}{scalar or vector, numeric times at which Uno's C will be calculated.}

\item{time}{string variable name, either time to event or end of follow-up.}

\item{event}{string variable name, indicator of whether \code{time} corresponds to an event (\code{event=1}) or censoring (\code{event=0}).}

\item{log_risk}{string variable name, log risk score predicted from model.}

\item{data}{dataframe, containing at least \code{time} and \code{event}.}
}
\value{
numeric AUC value
}
\description{
This function returns the KM-adjusted area under the receiver operating characteristic curves for survival models.
}
