% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unos_C.R
\name{unos_C}
\alias{unos_C}
\title{Uno's C statistic for survival models}
\usage{
unos_C(evaluate_at, time, event, log_risk, data)
}
\arguments{
\item{evaluate_at}{scalar or vector, numeric times at which Uno's C will be calculated.}

\item{time}{string variable name, either time to event or end of follow-up.}

\item{event}{string variable name, indicator of whether \code{time} corresponds to an event (\code{event=1}) or censoring (\code{event=0}).}

\item{log_risk}{string variable name, log risk score predicted from model.}

\item{data}{dataframe, containing at least \code{time} and \code{event}.}
}
\value{
A list with the following named slots:
\item{model_coeff}{dataframe with final model coefficients and standard error estimates (where applicable) for the analysis model.}
}
\description{
This function returns the time-varying and global Uno's C statistics for an input survival model's predictions.
}
